Phase 3 (partiel): troncatures pointeur, register, /W4
Conformement a roadmap.md Phase 3, avec un perimetre precise apres
discussion (voir roadmap.md pour le detail de la decision) :

- Suppression de 134 occurrences du mot-cle register (non standard
  depuis C++17) dans le code AzurVeda (Veda, VedaLibDemo, VedaLibImage,
  VedaMachineOGL, VedaLibSoundMP3/MP3SoundObject.cpp). Le code tiers
  vendu (fork de libmad, uniminixm) n'est pas touche, hors perimetre
  comme etabli en Phase 2.

- Correction des vraies troncatures pointeur -> entier 32 bits (bugs
  64-bit reels, pas de simples avertissements) avec
  intptr_t/uintptr_t/ptrdiff_t :
  - Veda/BaseType.cpp, VedaLibSoundMP3/MP3SoundObject.cpp :
    soustractions de pointeurs castees en int/unsigned int
  - VedaLibDemo/MarchinCubeSpace.cpp : arrondi d'alignement memoire
    sur 32 octets via un masque int -- corrompait entierement le
    pointeur en 64-bit (bug memoire reel, pas juste un avertissement)
  - VedaLibDemo/Object3DMarchCube.cpp : pointeur stocke dans un int
    puis re-additionne et recaste a trois endroits
  - VedaGUIWindowsMFC/LeftView.cpp : NMHDR::idFrom est UINT_PTR dans
    le SDK Win32, pas UINT -- la variable locale tronquait le handle
    de fenetre avant meme le cast (HWND) repere par le compilateur

- /W4 active projet entier (CMakeLists.txt racine). Etat des lieux :
  0 avertissement de troncature pointeur (C4311/C4312/C4302) restant
  dans le code AzurVeda ; 425 avertissements d'autres categories
  (C4244, C4100, C4189, C4996, signe/non-signe, etc.) + 72 dans le
  code tiers vendu, documentes dans roadmap.md/CHANGELOG.md comme
  travail futur plutot que traites maintenant (decision utilisateur :
  perimetre trop large pour ce passage -- plusieurs heures de travail
  avec risque de regressions comportementales sur du code de 2007
  jamais retexte). /WX non active tant que ce travail n'est pas fait.

Build complet revalide (cmake --preset ci-windows-x64
-DAZURVEDA_BUILD_MFC_EDITOR=ON && cmake --build --preset
ci-windows-x64-release), exit 0, aucune regression. Example01 et
Example02 executes pour verification runtime (Example02 tourne sa
boucle de rendu sans crash au demarrage, tue apres 5s par timeout).

+++ Veda/BaseType.cpp
@@ -6,6 +6,7 @@
 #include "BaseObject.h"
 #include "BaseContext.h"
 #include "PackString.h"
+#include <cstddef>
 #include <stdio.h>
 #endif
 
@@ -59,8 +60,8 @@ unsigned int BaseType::WriteFile(const char * _pFileName)
 		fwrite( pserializedForm, 1, serializedFormSize , ff );
 		fclose( ff );
 	}
-	int vtest = (int)pserializedFormAfter - (int)pserializedForm;
-	if( vtest != serializedFormSize )
+	ptrdiff_t vtest = pserializedFormAfter - pserializedForm;
+	if( static_cast<size_t>(vtest) != serializedFormSize )
 	{
 		vtest++;
 	}
